moving svg outline files
