+++ frontend/src/lib/components/Svg/outline/mailIcon.js
