+++ frontend/src/lib/components/Svg/outline/phoneIcon.js
