+++ frontend/src/lib/components/Svg/outline/rocket-launch.js
