+++ frontend/src/lib/components/Svg/outline/upload-icon.js
